requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
black==23.1.0
blinker==1.6
boto3==1.20.13
botocore==1.23.13
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exchangelib==5.0.1
fake-useragent==1.1.3
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==2.0.2
gspread==4.0.1
gspread-dataframe==3.3.0
gspread-pandas==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
hyperframe==6.0.1
idna==3.4
isodate==0.6.1
jmespath==0.10.0
kaitaistruct==0.10
Levenshtein==0.20.9
lxml==4.9.2
mypy-extensions==1.0.0
numpy
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.9
outcome==1.2.0
packaging==23.0
pandas

paramiko==2.11.0
parse==1.19.0
pathspec==0.11.1
platformdirs==3.2.0
protobuf==4.21.12
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydivert==2.1.0
pyee==8.2.2
Pygments==2.14.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
pyspnego==0.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-Levenshtein==0.20.9
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==2.13.7
requests==2.26.0
requests-html==0.10.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
selenium==4.8.3
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.0
thefuzz==0.19.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
w3lib==2.1.1
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
zipp==3.15.0
zstandard==0.20.0
termcolor